[refactor] controoler, service, repository test 코드 작성 시 custom 어노테이션 생성으로 중복되는 어노테이션 간소화
