+++ dongchimi-party-domain/src/main/java/com/dcm/global/annotation/RepositoryTest.java
@@ -1,12 +1,19 @@
-package com.dcm.common;
+package com.dcm.global.annotation;
 
 import com.dcm.global.config.JpaAuditingConfiguration;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
 @DataJpaTest
 @ActiveProfiles("test")
 @Import(JpaAuditingConfiguration.class)
-public class RepositoryTest {
+public @interface RepositoryTest {
 }
