+++ dongchimi-party-domain/src/main/java/com/dcm/global/annotation/ServiceTest.java
@@ -1,0 +1,16 @@
+package com.dcm.global.annotation;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@SpringBootTest
+@ActiveProfiles("test")
+public @interface ServiceTest {
+}
