--- dongchimi-party-domain/src/test/java/com/dcm/common/ControllerTest.java
@@ -1,10 +1,0 @@
-package com.dcm.common;
-
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.test.context.ActiveProfiles;
-
-@AutoConfigureRestDocs
-@ActiveProfiles("test")
-public class ControllerTest {
-
-}
