--- dongchimi-party-domain/src/test/java/com/dcm/common/ServiceTest.java
@@ -1,9 +1,0 @@
-package com.dcm.common;
-
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-public class ServiceTest {
-}
